[Update] Adjust Demo
