+++ Demo/JSONPreviewDemo/Core/ExampleJSON.swift
@@ -3,6 +3,7 @@
 //  JSONPreviewDemo
 //
 //  Created by Rakuyo on 2024/1/8.
+//  Copyright © 2024 Rakuyo. All rights reserved.
 //
 
 import Foundation
