+++ Demo/JSONPreviewDemo/Core/List/ListTableViewCell.swift
@@ -1,0 +1,19 @@
+//
+//  ListTableViewCell.swift
+//  JSONPreviewDemo
+//
+//  Created by Rakuyo on 2024/3/25.
+//  Copyright © 2024 Rakuyo. All rights reserved.
+//
+
+import UIKit
+
+final class ListTableViewCell: UITableViewCell {
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: .subtitle, reuseIdentifier: reuseIdentifier)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
